TGP-2894: moved files into packages
