+++ app/forms/goodsProfile/RemoveGoodsRecordFormProvider.scala
@@ -14,32 +14,17 @@
  * limitations under the License.
  */
 
-package forms
+package forms.goodsProfile
 
-import forms.behaviours.BooleanFieldBehaviours
-import play.api.data.FormError
+import forms.mappings.Mappings
+import play.api.data.Form
 
-class RemoveGoodsRecordFormProviderSpec extends BooleanFieldBehaviours {
+import javax.inject.Inject
 
-  val requiredKey = "removeGoodsRecord.error.required"
-  val invalidKey  = "error.boolean"
+class RemoveGoodsRecordFormProvider @Inject() extends Mappings {
 
-  val form = new RemoveGoodsRecordFormProvider()()
-
-  ".value" - {
-
-    val fieldName = "value"
-
-    behave like booleanField(
-      form,
-      fieldName,
-      invalidError = FormError(fieldName, invalidKey)
+  def apply(): Form[Boolean] =
+    Form(
+      "value" -> boolean("removeGoodsRecord.error.required")
     )
-
-    behave like mandatoryField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey)
-    )
-  }
 }
