+++ app/views/goodsProfile/GoodsRecordsSearchResultEmptyView.scala.html
