+++ app/views/goodsProfile/GoodsRecordsSearchResultView.scala.html
