+++ app/views/goodsProfile/GoodsRecordsView.scala.html
