+++ test/forms/goodsProfile/RemoveGoodsRecordFormProviderSpec.scala
@@ -14,17 +14,32 @@
  * limitations under the License.
  */
 
-package forms
+package forms.goodsProfile
 
-import javax.inject.Inject
+import forms.behaviours.BooleanFieldBehaviours
+import play.api.data.FormError
 
-import forms.mappings.Mappings
-import play.api.data.Form
+class RemoveGoodsRecordFormProviderSpec extends BooleanFieldBehaviours {
 
-class RemoveGoodsRecordFormProvider @Inject() extends Mappings {
+  val requiredKey = "removeGoodsRecord.error.required"
+  val invalidKey  = "error.boolean"
 
-  def apply(): Form[Boolean] =
-    Form(
-      "value" -> boolean("removeGoodsRecord.error.required")
+  val form = new RemoveGoodsRecordFormProvider()()
+
+  ".value" - {
+
+    val fieldName = "value"
+
+    behave like booleanField(
+      form,
+      fieldName,
+      invalidError = FormError(fieldName, invalidKey)
     )
+
+    behave like mandatoryField(
+      form,
+      fieldName,
+      requiredError = FormError(fieldName, requiredKey)
+    )
+  }
 }
